lots of updates. use sgDevice, lots of updates to dependencies.
An experimental setup.
There might still be bugs, crashes and design flaws, but it seems to
work as a proof of concept.
Message bursts still lead to io-errors in pd, which produce audible "clicks".
Splitting a patch into "io" and an "audio" part can help:

- The io patch receives messages (from midi, keyboard,...),
does message processing and communicates with the actual "audio" patch
via TCP or UDP utilising pds "netsend" and "netreceive" objects
- The "audio" patch receives messages like "play note with pitch x" does
the audio processing

The `./scripts/run.fish` supports this model now.